Refactor C# AutomationObject
