+++ src/VisualStudio/VisualBasic/Impl/Microsoft.VisualStudio.LanguageServices.VisualBasic.vbproj
@@ -14,14 +14,8 @@
     <DeployExtension>false</DeployExtension>
   </PropertyGroup>
   <ItemGroup Label="PkgDef">
-    <PkgDefInstalledProduct Include="{574fc912-f74f-4b4e-92c3-f695c208a2bb}"
-                            Name="Microsoft Visual Basic"
-                            DisplayName="#113"
-                            ProductDetails="#114" />
-    <PkgDefPackageRegistration Include="{574fc912-f74f-4b4e-92c3-f695c208a2bb}"
-                               Name="VisualBasicPackage"
-                               Class="Microsoft.VisualStudio.LanguageServices.VisualBasic.VisualBasicPackage"
-                               AllowsBackgroundLoad="true" />
+    <PkgDefInstalledProduct Include="{574fc912-f74f-4b4e-92c3-f695c208a2bb}" Name="Microsoft Visual Basic" DisplayName="#113" ProductDetails="#114" />
+    <PkgDefPackageRegistration Include="{574fc912-f74f-4b4e-92c3-f695c208a2bb}" Name="VisualBasicPackage" Class="Microsoft.VisualStudio.LanguageServices.VisualBasic.VisualBasicPackage" AllowsBackgroundLoad="true" />
     <None Include="PackageRegistration.pkgdef" PkgDefEntry="FileContent" />
   </ItemGroup>
   <ItemGroup Label="Project References">
